requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
Flask_Session==0.8.0
itsdangerous==2.2.0
twilio==9.2.3
WTForms==3.1.2
Werkzeug==3.0.3
gunicorn==23.0.0
Flask-Login==0.6.3
email_validator==2.2.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
# google-cloud-vision==3.7.4
azure-cognitiveservices-vision-computervision==0.9.0
flask-debugtoolbar==0.15.1
rich==13.8.0